+++ src/settings/AppearanceSettingsViewer.ts
@@ -9,31 +9,11 @@ import { TimeFormat, WeekStart } from "../api/common/TutanotaConstants"
 import { downcast, incrementDate, noOp, promiseMap } from "@tutao/tutanota-utils"
 import { UserSettingsGroupRootTypeRef } from "../api/entities/tutanota/TypeRefs.js"
 import { getHourCycle } from "../misc/Formatter"
-import type { ThemeId, ThemePreference } from "../gui/theme"
-import { themeController } from "../gui/theme"
+import { themeController, ThemeId, themeOptions, ThemePreference } from "../gui/theme"
 import type { UpdatableSettingsViewer } from "./SettingsView"
 import { isDesktop } from "../api/common/Env"
 import { locator } from "../api/main/MainLocator"
 import { EntityUpdateData, isUpdateForTypeRef } from "../api/common/utils/EntityUpdateUtils.js"
-
-export const themeOptions = [
-	{
-		name: "systemThemePref_label",
-		value: "auto:light|dark",
-	},
-	{
-		name: "light_label",
-		value: "light",
-	},
-	{
-		name: "dark_label",
-		value: "dark",
-	},
-	{
-		name: "blue_label",
-		value: "blue",
-	},
-] as const
 
 export class AppearanceSettingsViewer implements UpdatableSettingsViewer {
 	private _customThemes: Array<ThemeId> | null = null
